+++ defmt/src/export/mod.rs
@@ -36,31 +36,39 @@ pub fn fetch_add_string_index() -> u16 {
 /// Get and clear the logged bytes
 #[cfg(feature = "unstable-test")]
 pub fn fetch_bytes() -> Vec<u8> {
-    BYTES.with(|b| core::mem::replace(&mut *b.borrow_mut(), Vec::new()))
+    BYTES.with(|b| core::mem::take(&mut *b.borrow_mut()))
 }
 
+/// Only to be used by the defmt macros
+/// Safety: must be paired with a later call to release()
 #[cfg(feature = "unstable-test")]
-pub fn acquire() {}
+pub unsafe fn acquire() {}
 
+/// Only to be used by the defmt macros
+/// Safety: must be paired with a later call to release()
 #[cfg(not(feature = "unstable-test"))]
 #[inline(always)]
-pub fn acquire() {
+pub unsafe fn acquire() {
     extern "Rust" {
         fn _defmt_acquire();
     }
-    unsafe { _defmt_acquire() }
+    _defmt_acquire()
 }
 
+/// Only to be used by the defmt macros
+/// Safety: must follow an earlier call to acquire()
 #[cfg(feature = "unstable-test")]
-pub fn release() {}
+pub unsafe fn release() {}
 
+/// Only to be used by the defmt macros
+/// Safety: must follow an earlier call to acquire()
 #[cfg(not(feature = "unstable-test"))]
 #[inline(always)]
-pub fn release() {
+pub unsafe fn release() {
     extern "Rust" {
         fn _defmt_release();
     }
-    unsafe { _defmt_release() }
+    _defmt_release()
 }
 
 #[cfg(feature = "unstable-test")]
@@ -190,13 +198,13 @@ pub fn bool(b: &bool) {
 
 /// Implementation detail
 pub fn debug(val: &dyn core::fmt::Debug) {
-    core::write!(FmtWrite, "{:?}", val).ok();
+    core::write!(FmtWrite, "{val:?}").ok();
     write(&[0xff]);
 }
 
 /// Implementation detail
 pub fn display(val: &dyn core::fmt::Display) {
-    core::write!(FmtWrite, "{}", val).ok();
+    core::write!(FmtWrite, "{val}").ok();
     write(&[0xff]);
 }